advanced_ai_agents/multi_agent_apps/agent_teams/ai_vc_due_diligence_agent_team/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
loguru>=0.7.0
gunicorn>=21.0.0
google-adk>=1.0.0
google-genai>=1.0.0
matplotlib>=3.8.0
